don't consider disc number when deciding if two tracks are the same
It's fairly common in cases of multiple file extensions to get cases
where the disc of one is '1' and the other is empty string, which then
ends up in both getting unique track IDs, and the extension priority
mechanism not working.

Disc number shouldn't be considered to be a part of a track's identity.